noteId: "2828347064af11eb8191153eb319e31b" tags: []

MinimaxAdaptiveControl.K — Method

K(mac::MAController)

Returns the feedback gain such that u = -Kx

MinimaxAdaptiveControl.Tsyn — Method

T(mac, model, tol = 0.01) 
synthesize a common T using convex programming such that
inequalities (19) and (20) are fulfilled.

...

Arguments:

mac::MAController MinimaxAdaptiveControl controller object

model::JuMP.Model A user supplied JuMP model. Currently the solvers Mosek and Hypatia works.

tol::Real = 0.01 Forcing (19) and (20) to hold with margin, i.e. T - tol*I >= ...

...

MinimaxAdaptiveControl.X — Method

X(mac,T,i, k, tol) where P<:Number

Synthesize a symmetric matrix X such that inequality (19) is satisfied iff X is positive semidefinite. ...

Arguments:

mac::MAController MinimaxAdaptiveControl controller object

T::AbstractMatrix Either a matrix or a JuMP variable

i::Integer Index variable,

k::Integer Index Variable

tol::Real = 0.01 Forcing (19) to hold with margin, i.e. T - tol*I >= ...

...

MinimaxAdaptiveControl.Z — Method

Z(mac::MAController{P},
T::Symmetric{VariableRef,Array{VariableRef,2}}, 
i::Integer, j::Integer, k::Integer) where P<:Number

Synthesize a symmetric matrix Z such that inequality (20) is satisfied iff Z is positive semidefinite.

...

Arguments:

mac::MAController MinimaxAdaptiveControl controller object

T::AbstractMatrix Either a matrix or a JuMP variable

i::Integer Index variable

j::Integer Index variable

k::Integer Index Variable

tol::Real = 0.01 Forcing (20) to hold with margin, i.e. T - tol*I >= ...

...

MinimaxAdaptiveControl.dare — Method

dare(A, B, Q, R)

Compute X, the solution to the discrete-time algebraic Riccati equation, defined as A'XA - X - (A'XB)(B'XB + R)^-1(B'XA) + Q = 0, where Q>=0.

This version relaxes the requirement that R>0, rather R = [R1 0;0 R2] where R1 >0, R2 < 0. This is a usuful formulation for H_∞ synthesis.

Algorithm taken from: Laub, "A Schur Method for Solving Algebraic Riccati Equations." Relaxes the positive definiteness of R, useful for robust control.

http://dspace.mit.edu/bitstream/handle/1721.1/1301/R-0859-05666488.pdf

Implementation stolen from: ControlSystems.jl

MinimaxAdaptiveControl.update! — Method

update!(mac,x, u)

Update the internal states of the controller based on current state x and control signal u. ...

Arguments:

mac::MAController MinimaxAdaptiveControl controller object

x::AbstractArray{T,1} The current state

u::Abstractarray{T,1} The previous control signal

...

MinimaxAdaptiveControl.Candidate — Type

Candidate{T<:Number}

Fields:

A::AbstractMatrix{T} System matrix

B::AbstractMatrix{T} Input Gain matrix

K::AbstractMatrix{T} $H_\infty$` feedback gain

P::AbstractMatrix{T} Stationary solution to the Riccati equation

hist::Base.RefValue{<:Real} History, $\sum _{k=0}^N \lambda^{N-k}|x_{k+1} - Ax_k -Bu_k|^2$

lam::T Forgetting factor $\lambda$

MinimaxAdaptiveControl.MAController — Method

mac = MAController(As, Bs, Q, R, γ, x0, lam)

Create a MAController (minimax adaptive controller) mac::MAController{T} with matrices containing elements of type T, solving \min_u\max_w,i\sum (|xt|^2\Q + |ut|^2\R) - \gamma^2 \sum |w_t|^2 ...

Arguments

As::AbstractVector{<:AbstractMatrix{T}} A vector of state matrices

Bs::AbstractVector{<:AbstractMatrix{T}} A vector of input gains

Q::AbstractMatrix, State penalty matrix |x|_Q^2

R::AbstractMatrix, Control penalty matrix |u|_R^2

γ::Real, Disturbance penalty, \H_\infty gain

x0::AbstractVector{<:Number} Initial state

lam::Real=1 Forgetting factor

...