Main.MinimaxAdaptiveControl.K — Method

K(mac::MAController)

Returns the feedback gain such that u = -Kx

Main.MinimaxAdaptiveControl.Pcommon — Method

Pcommon(mac::MAController) synthesizes a common P such that

TO IMPLEMENT

P >= Q + K'RK + (A-BK)(P^(-1) - 1/gamma^2I)^(-1)(A-BK) for all candidates (A,B,K).

Main.MinimaxAdaptiveControl.T — Method

T(mac::MAController) synthesizes a common T using convex
programming such that

TO IMPLEMENT!

T >= Q + Kj'RKj + (Ai-BiKj)(P^(-1) - 1/gamma^2I)^(-1)(Ai-BiKj) and T >= Q + Kk'RKk ...

Main.MinimaxAdaptiveControl.dare — Method

dare(A, B, Q, R)

Compute X, the solution to the discrete-time algebraic Riccati equation, defined as A'XA - X - (A'XB)(B'XB + R)^-1(B'XA) + Q = 0, where Q>=0.

This version relaxes the requirement that R>0, rather R = [R1 0;0 R2] where R1 >0, R2 < 0. This is a usuful formulation for H_∞ synthesis.

Algorithm taken from: Laub, "A Schur Method for Solving Algebraic Riccati Equations." Relaxes the positive definiteness of R, useful for robust control.

http://dspace.mit.edu/bitstream/handle/1721.1/1301/R-0859-05666488.pdf

Implementation stolen from: ControlSystems.jl

Main.MinimaxAdaptiveControl.update! — Method

update!(mac::MAController, 
    x::AbstractArray{T,1}, u::AbstractArray{T,1}) where T<:Number

Update the internal states of the controller based on current state x and control signal u.

Main.MinimaxAdaptiveControl.Candidate — Type

Candidate{T<:Number}

Fields:

A::AbstractMatrix{T} System matrix

B::AbstractMatrix{T} Input Gain matrix

K::AbstractMatrix{T} $H_\infty$` feedback gain

P::AbstractMatrix{T} Stationary solution to the Riccati equation

hist::Base.RefValue{<:Real} History, $\sum _{k=0}^N \lambda^{N-k}|x_{k+1} - Ax_k -Bu_k|^2$

lam::T Forgetting factor $\lambda$